% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{co_vol}
\alias{co_vol}
\title{Work out the comoving volume for a given redshift or array of redshifts}
\usage{
co_vol(redshift, o_m, o_k, o_l, h_0)
}
\arguments{
\item{redshift}{either a single redshift or an array of redshifts.}

\item{h_0}{H0 = 100 * h}

\item{omega_m}{Mass density (often 0.3 in LCDM)}

\item{omega_k}{Effective mass density of relativistic particles (often 0. in LCDM)}

\item{omega_l}{Effective mass density of dark energy (often 0.7 in LCDM)}
}
\description{
Work out the comoving volume for a given redshift or array of redshifts
}
