% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{comoving_trans_distance}
\alias{comoving_trans_distance}
\title{Comoving transverse distance.}
\usage{
comoving_trans_distance(redshift, omega_m, omega_k, omega_l, h0)
}
\description{
Comoving transverse distance.
}
