% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{comoving_transverse_distances}
\alias{comoving_transverse_distances}
\title{Comoving transverse distances for an array}
\usage{
comoving_transverse_distances(redshifts, omega_m, omega_k, omega_l, h0)
}
\description{
Comoving transverse distances for an array
}
