% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{universe_age}
\alias{universe_age}
\title{Unviverse age at a particular redshift in Gyr}
\usage{
universe_age(redshift, omega_m, omega_k, omega_l, h0)
}
\description{
Unviverse age at a particular redshift in Gyr
}
