% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{universe_ages}
\alias{universe_ages}
\title{Universe ages at multiple redshifts}
\usage{
universe_ages(redshifts, omega_m, omega_k, omega_l, h0)
}
\description{
Universe ages at multiple redshifts
}
