% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{z_at}
\alias{z_at}
\title{Do the approximate inverse of the cosmological function}
\usage{
z_at(cosmo_function, value, omega_m, omega_k, omega_l, h0)
}
\arguments{
\item{value}{The value of the function that you wan't to find the redshift for}

\item{como_function}{The cosmological function to inverse}
}
\value{
The redshift at the given value
}
\description{
Do the approximate inverse of the cosmological function
}
